+++ src/components/datepicker/DatePickerInputRange.js
@@ -212,7 +212,7 @@ class DatePickerInputRange extends PureComponent {
   };
 
   render() {
-    const { bold, disabled, error, helpText, inverse, readOnly, size, ...others } = this.props;
+    const { bold, disabled, error, helpText, inverse, readOnly, size, warning, width, ...others } = this.props;
 
     const boxProps = pickBoxProps(others);
 
@@ -223,15 +223,16 @@ class DatePickerInputRange extends PureComponent {
       [theme['is-read-only']]: readOnly,
       [theme['has-error']]: error,
       [theme['has-focus']]: this.state.inputHasFocus,
+      [theme['has-warning']]: warning,
     });
 
     return (
       <Box className={classNames} {...boxProps}>
-        <div className={theme['input-wrapper']}>
+        <div className={theme['input-wrapper']} style={{ width }}>
           {this.renderIcon()}
           {this.renderDayPickerInput()}
         </div>
-        <ValidationText error={error} help={helpText} inverse={inverse} />
+        <ValidationText error={error} help={helpText} inverse={inverse} warning={warning} />
       </Box>
     );
   }
@@ -268,6 +269,10 @@ DatePickerInputRange.propTypes = {
   size: PropTypes.oneOf(['small', 'medium', 'large']),
   inputStartDateValue: PropTypes.oneOfType([PropTypes.string, PropTypes.instanceOf(Date)]),
   inputEndDateValue: PropTypes.oneOfType([PropTypes.string, PropTypes.instanceOf(Date)]),
+  /** The text to use as warning message below the input. */
+  warning: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
+  /** A custom width for the input field */
+  width: PropTypes.string,
 };
 
 DatePickerInputRange.defaultProps = {
@@ -276,6 +281,7 @@ DatePickerInputRange.defaultProps = {
   inverse: false,
   readOnly: false,
   size: 'medium',
+  width: '210px',
 };
 
 export default DatePickerInputRange;