+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 ### Fixed
 
 - `DatePickerInput`: merged the passed inputProps with the internal inputProps. ([@duivvv](https://github.com/duivvv) in [#503](https://github.com/teamleadercrm/ui/pull/503))
+- `Popover`: fixed buggy flexbox behavior. ([@driesd](https://github.com/driesd) in [#508](https://github.com/teamleadercrm/ui/pull/508))
 
 ## [0.19.8] - 2019-01-15